Major cleanup
The sweet spot for us seems to be the *Analyze* phase so all the
machinery (scanning the AST and symbols) was moved here and for now we
are interested only in switch and variable nodes in AST. Also I figured
out how to accumulate the AST scan result because I am smart.

A lot of things were removed, mostrly finddling around garbage.