Refactoring in KoguListener
The notion of State was introduced. It contains four sets:
  - the set of Enum declarations for the compilation unit
  - the set of all Enum definitions available at *ANALYZE* phase
  - the set of switch statements over Enums per compilation unit
  - the set of all partial switches. These are accumulated through compilation

Notions for Enum definitions and declarations as well as switches over
them were introduced. Current flow goes like this: for each compilation
unit, rinse the state, scan the AST and the symbols table, update the
state and evaluate it. The result of the evaluation process is a Report,
containing stuff yet to be decided.