Working prototype for **Enum**-s defined in local package
Enum declarations have been removed because frankly we do not
care what is the name of the **Enum** variable or where it was
defined.

We still do not handle cases, where the **Enum** is imported from
a third-party jar, since we resolve **Enum**-s defined in the local
package only. This should be fixed by looking them up in
*namedImportScope*
and *starImportScope*

Evaluation report got slightly enhaced, thou it seems that the flow
of report generation is broken, since currently a new *Report* will
be created for each compilation unit, whereas it would be nice to
have a single report for the full compilation. Also the file, line
and column of a partial switch should be available in the report.