Implemented fully functional **Enum** resolution
Now we are able to resolve **Enum**-s in any package, inside of
nested classes of arbitrary depth and in imports in named and
starred scopes of third party libraries.

It turns out the *javac* API for Java 11 differs from that of Java 8.
This implementation works for Java 11. With some modifications it also
works for Java 8. This calls for separate branches.

Also the report should probably use a **Set** of partial switches,
not a **Map**. This will we changes once I figure out how to get the
location of the switch statement (file : line : character).