Error reporting implemented
Implemented the error reporting for Kogu. I gan't understand how to also
include the file name and switch statement position in the source
but other than that it works, providing a list of missing **Enum**
members in the error message.

The **Report** type has been removed since now we report a all partial
**Enum**-s after the evaluation of each compilation unit.